# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/building": {
            "get": {
                "tags": [
                    "Building"
                ],
                "description": "Get list of all buildings",
                "operationId": "3414f12e1d87ae6869b3c8173181fc3d",
                "responses": {
                    "200": {
                        "description": "List of buildings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BuildingDTO"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Building"
                ],
                "description": "Create a building",
                "operationId": "81585b0eb585c93578fd2c9806339bb6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBuildingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Building created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "409": {
                        "description": "Id already taken"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
...
        "/api/building/geo/rect": {
            "post": {
                "tags": [
                    "Building"
                ],
                "description": "Find buildings inside a rect area",
                "operationId": "c9694d843951c36ecc389f75e505537e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Rect"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of buildings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BuildingDTO"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
        "/api/building/geo/circle": {
            "post": {
                "tags": [
                    "Building"
                ],
                "description": "Find buildings inside a circle area",
                "operationId": "5c6eec6938e5b6ec1cfa1d60029f06b5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Circle"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of buildings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BuildingDTO"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
        "/api/building/{id}": {
            "get": {
                "tags": [
                    "Building"
                ],
                "description": "Get info about specific building",
                "operationId": "954726f8d6bc5842c83d324fadba43b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Building",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BuildingDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cannot find building by id"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Building"
                ],
                "description": "Update building",
                "operationId": "666c6b0d6d95f3f6e3baa86f4bb02605",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Building updated"
                    },
                    "404": {
                        "description": "Cannot find building by id"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Building"
                ],
                "description": "Delete building",
                "operationId": "93415d94ec193f94d107514d428dca8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Building updated"
                    },
                    "404": {
                        "description": "Cannot find building by id"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
        "/api/building/{id}/companies": {
            "get": {
                "tags": [
                    "Building"
                ],
                "description": "List companies in building",
                "operationId": "f492ca07d0c6eeab76c7d55b78924f73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of companies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CompanyDTO"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cannot find building by id"
                    }
                },
                "security": [
                    {
                        "http": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BuildingDTO": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "coordinates": {
                        "$ref": "#/components/schemas/Point"
                    }
                },
                "type": "object"
            },
            "CompanyDTO": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "buildingId": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CreateBuildingRequest": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "address": {
                        "type": "address"
                    },
                    "coordinates": {
                        "$ref": "#/components/schemas/Point"
                    }
                },
                "type": "object"
            },
...
            "Circle": {
                "properties": {
                    "center": {
                        "$ref": "#/components/schemas/Point"
                    },
                    "radius": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "Point": {
                "properties": {
                    "latitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "Rect": {
                "properties": {
                    "topLeft": {
                        "$ref": "#/components/schemas/Point"
                    },
                    "bottomRight": {
                        "$ref": "#/components/schemas/Point"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "http": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}